requirements.txt
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
schedule>=1.2.0
python-dotenv>=1.0.0
skyfield>=1.42
astropy>=5.3.0
sqlite3
ephem>=4.1

folium>=0.12.0

twilio>=8.0.0
python-telegram-bot>=20.0

apscheduler>=3.9.0

timezonefinder>=6.0.0
